+++ apps/social/serializers.py
@@ -162,6 +162,7 @@ class ListMissionSerializer(
             "description",
             "attachment",
             "attachment_type",
+            "thumbnail",
             "is_completed",
             "completed_info",
         ]