+++ tests/unit/social/test_social_serializers.py
@@ -277,6 +277,7 @@ class TestListMissionSerializer:
             "description",
             "attachment",
             "attachment_type",
+            "thumbnail",
             "is_completed",
             "completed_info",
         ]